split template helpers and slug utils out of markdown.go
